+++ tests/pstl-tests/invalid_vector_iterators.cpp
@@ -112,7 +112,7 @@ TEST_F(InvalidIterators, TestCountIf1) {
   int n_elems = 128;
 
   for (int i = 0; i < n_elems; i++) {
-    float x = ((float)std::rand()) / RAND_MAX;
+    float x = (((float)std::rand()) / (float)RAND_MAX);
     v.push_back(x);
   }
 
@@ -129,7 +129,7 @@ TEST_F(InvalidIterators, TestCountIf2) {
   int n_elems = 128;
 
   for (int i = 0; i < n_elems; i++) {
-    float x = ((float)std::rand()) / RAND_MAX;
+    float x = (((float)std::rand()) / (float)RAND_MAX);
     v.push_back(x);
   }
 
@@ -146,7 +146,7 @@ TEST_F(InvalidIterators, TestCountIf3) {
   int n_elems = 128;
 
   for (int i = 0; i < n_elems; i++) {
-    float x = ((float)std::rand()) / RAND_MAX;
+    float x = (((float)std::rand()) / (float)RAND_MAX);
     v.push_back(x);
   }
